Add infrastructure implementation for browser
