+++ packages/browser/package.json
@@ -35,6 +35,7 @@
     "eslint-config-prettier": "^4.1.0",
     "jest": "^24.7.1",
     "prettier": "^1.17.0",
+    "ts-jest": "^24.0.2",
     "typescript": "^3.4.5"
   },
   "dependencies": {
@@ -43,8 +44,6 @@
     "firebase": "^5.10.1",
     "inversify": "^5.0.1",
     "lodash": "^4.17.11",
-    "reflect-metadata": "^0.1.13",
-    "ts-jest": "^24.0.2",
-    "typescript": "^3.4.5"
+    "reflect-metadata": "^0.1.13"
   }
 }
