+++ packages/browser/src/Firebase/index.ts
@@ -1,0 +1,1 @@
+export * from "./symbols"
