+++ packages/browser/src/Firebase/symbols.ts
@@ -1,0 +1,6 @@
+export const symbols = {
+  firebaseApp: Symbol.for("firebaseApp"),
+  firebaseAuthProvider: Symbol.for("firebaseAuthProvider"),
+  firebaseAuth: Symbol.for("firebaseAuth"),
+  firestore: Symbol.for("firestore")
+};
