+++ packages/browser/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./Auth/index"
-export * from "./Firebase/index"
